feat: styling review
